refactor(spark): share the Spark 3.x legacy parquet file format via a common base
